// frontend/lib/api/generated/model/ordenCompraDetalleIn.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { OrdenCompraDistribucionIn } from './ordenCompraDistribucionIn';

export interface OrdenCompraDetalleIn {
  producto_contratado_id: number;
  /** @exclusiveMinimum 0 */
  cantidad_solicitada: number;
  /** @minItems 1 */
  distribucion: OrdenCompraDistribucionIn[];
}